requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
pandas>=1.1.3
scikit-learn>=0.23.2
matplotlib>=3.3.2
seaborn>=0.11.0
nltk>=3.5
pillow>=7.2.0
tqdm>=4.50.2
lime>=0.2.0
pyyaml>=5.3.1
requests>=2.24.0
tensorflow-hub>=0.10.0
